Do not rely on hardcoded NitroAdjudicator address
On every testground run:
1. Testground instance with sequence number 1 checks if NitroAdjudicator
has been deployed. If not, the instance deploys the adjudicator.
2. All instances block until the adjudicator is deployed.

Create2 is used for deploys, so a new deploy is only triggered when the
creation code for NitroAdjudicator changes. Since NitroAdjudicator has no
constructor, the creation code is the same as the bytecode.

Note that the following are important:
- Create2Deployer must exist at the hardcoded address.
- The same nonce should be used for all deploys.

The benefit of the above approach is that the testsuite will automatically
use a deployed NitroAdjudicator contract version from the go-nitro package. Before
this work, NitroAdjudicator version deployed on the chain could differ
from the version supplied by the go-nitro package.